Tweaked the zoom behavior to better handle video larger than the available display area
This fixes a bug where it was impossible to zoom in at a reasonable size on a specific area of a video with substantially larger dimensions than the available screen size. Basically you could zoom but it was always anchored in the upper left corner instead of honoring the zoom center.

I also made it so that resizing the window will re-render the frame (recentering the zoom if necessary.)